refactoring logic for fetching cocktail cards
